give the source code an organized structure for its folder
