+++ .env-sample
@@ -1,1 +1,2 @@
 MONGO_URL='your url here'
+MONGO_DB_NAME='defaults to test'
