+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "week-10",
+  "name": "dynamic-api",
   "version": "1.0.0",
   "description": "",
   "main": "./src/index.js",
