+++ src/app/app.js
@@ -1,0 +1,6 @@
+const express = require("express");
+
+// defining the Express app. Please don't add any additional code. See ./README
+const app = express();
+
+module.exports = app;
