+++ src/app/index.js
@@ -1,0 +1,3 @@
+const app = require("./app.routes");
+
+module.exports = app;
