+++ src/index.js
@@ -1,8 +1,10 @@
 // This file is: ./src/index.js
+const app = require('./app')
 
 //importing the dependencies
-const { app, startDatabase } = require("./app-common.js");
-const { request } = require("express");
+
+const {  startDatabase } = require("./database/sys.functions");
+
 
 // support production deployment on a port configured on the hosting server
 // default to the dev port number otherwise
@@ -13,28 +15,11 @@ const MONGO_URL = process.env.MONGO_URL;
 // connect to our database then start the web server
 // https://www.mongodb.com/
 if (MONGO_URL) {
-  // endpoint to return top level api
-  // much like a switch statement
-  app.get("/", async (req, res) => {
-    res.send({
-      message: "Storefront API. See documentation for use.",
-    });
-  });
-
-  app.use("/products", require("./routes/productsRoutes"));
-  app.use("/logos", require("./routes/logosRoutes"));
-  app.use("/stores", require("./routes/storesRoutes"));
-  app.use("/categories", require("./routes/categoriesRoutes"));
-  app.use("/product-types", require("./routes/product-typesRoutes"));
-  app.use("/variations", require("./routes/variationsRoutes"));
-
-  // new routes
-  app.use("/api/tasks", require("./routes/tasksRoutes"));
 
   startDatabase().then(async () => {
     // `then` start the web server after the database starts
     app.listen(port, async () => {
-      console.log(`Web server has started on port ${port}`);
+      console.log(`Web server has started on port http://localhost:${port}`);
     });
   });
 } else {
